feat!: Replace shell based hooks with Python based hooks.
Renames Python hooks from `slurm` to `slurmhelpers`. This way it is more obvious that these hooks are for augmenting the functionality of slurm, not a custom rebuild of slurm.

BREAKING CHANGE: Old hook behavior is now changed to instead work with Python using `snap-helpers`. Now the hooks can be more advanced.